+++ PawBnB/src/Views/SignUpSitters/SignUpSitters.jsx
@@ -1,7 +1,9 @@
 import React from "react";
 
 const SignUpSitters = () => {
-  return <div>SignUpSitters</div>;
+  return (
+    <Formulario text= 'Registrate como Cuidador' />
+  )
 };
 
 export default SignUpSitters;
